==> 1. Data Cleaning fashion_dataset.ipynb <==
# Data Analysis

>Importing libraries

import pandas as pd
import numpy as np
import pandasql as ps
from pandasql import sqldf
import ast
import random
import matplotlib.pyplot as plt
import seaborn as sns
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import BayesianRidge
from sklearn import preprocessing
from sklearn import tree
from sklearn.tree import DecisionTreeClassifier
from sklearn import metrics
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import RandomForestClassifier

fashion_dataset= pd.read_csv(r".\fashion dataset.csv")
brand_details = pd.read_excel(r".\fashion brand details.xlsx")

# Data Cleaning & Transformation

>Importing and viewing the fashion dataset


fashion_dataset= pd.read_csv(r".\fashion dataset.csv")
fashion_dataset

>Finding the number of unique brands in the fashion dataset

len(pd.unique(fashion_dataset['brand']))

>Importing the brand details dataset

brand_details = pd.read_excel(r".\fashion brand details.xlsx")
brand_details

>Counting the number of unique brand names in the brand details dataset

len(pd.unique(brand_details['brand_name']))

>Finding the number of null and duplicated values in each dataset

brand_details.isna().sum()

brand_details.duplicated().sum()

fashion_dataset.isna().sum()

fashion_dataset.duplicated().sum()

>Testing out duplicate dropping

df_dupl = fashion_dataset.drop_duplicates(keep = "last")
len(pd.unique(df_dupl["brand"]))

>Dropping duplicates in fashion dataset

fashion_dataset.drop_duplicates(inplace = True, keep = "last")
fashion_dataset

>Dropping duplicate product names and IDs

fashion_dataset.drop_duplicates(subset=['name','p_id'], inplace=True)

fashion_dataset.info()

> Expanding the p_attributes column


fashion_dataset['p_attributes'] = fashion_dataset['p_attributes']
not_null = pd.notnull(fashion_dataset['p_attributes'])
fashion2 = fashion_dataset[not_null]['p_attributes'].apply(lambda x: ast.literal_eval(x))
temp = pd.DataFrame([*fashion2], fashion2.index)
fashion_dataset = pd.concat([fashion_dataset, temp], axis=1)

>Viewing all columns of the dataset

print(list(fashion_dataset.columns))

>Renaming columns for use in SQL and MapReduce

fashion_dataset.rename(columns={'Where-to-wear':'Where_to_wear','Dupatta Pattern':'Dupatta_Pattern', 'Kurta Pattern':'Kurta_Pattern','Bottom Pattern': 'Bottom_Pattern', 'Top Pattern': 'Top_Pattern','Print or Pattern Type': 'Print_Pattern_Type', 'Knit or Woven':'Knit_or_Woven', 'Weave Type':'Weave_Type', 'Weave Pattern':'Weave_Pattern','Dupatta Fabric':'Dupatta_Fabric','Bottom Fabric':'Bottom_Fabric', 'Top Fabric':'Top_Fabric','Fabric Purity':'Fabric_Purity', 'Fabric Type':'Fabric_Type','Blouse Fabric':'Blouse_Fabric', 'Better Cotton Initiative':'Better_Cotton', 'Saree Fabric':'Saree_Fabric', 'Colour Family':'Colour_Family'}, inplace=True)

fashion_dataset['Kurta_Pattern'].isna().sum()

fashion_dataset['Weave_Type'].isna().sum()

fashion_dataset['Knit_or_Woven'].isna().sum()

fashion_dataset['Weave_Pattern'].isna().sum()

### Filling the Fabric column

pd.unique(fashion_dataset['Fabric'])

sqldf("select count(Fabric) from fashion_dataset where Fabric = 'NA'")

fashion_dataset['Fabric'].isna().sum()

pd.unique(fashion_dataset['Bottom_Fabric'])

fashion_dataset['Bottom_Fabric'] = fashion_dataset['Bottom_Fabric'].where(fashion_dataset['Bottom_Fabric'] != 'NA', None)

pd.unique(fashion_dataset['Blouse_Fabric'])

fashion_dataset['Blouse_Fabric'] = fashion_dataset['Blouse_Fabric'].where(fashion_dataset['Blouse_Fabric'] != 'NA', None)

pd.unique(fashion_dataset['Top_Fabric'])

pd.unique(fashion_dataset['Fabric_Type'])

sqldf("select Better_Cotton, Fabric from fashion_dataset where Fabric is not null and Better_Cotton is not null")

fashion_dataset['Fabric_Type'] = fashion_dataset['Fabric_Type'].where(fashion_dataset['Fabric_Type'] != 'NA', None)

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['Top_Fabric'])

fashion_dataset['Fabric'].isna().sum()

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['Bottom_Fabric'])

fashion_dataset['Fabric'].isna().sum()

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['Fabric_Type'])

fashion_dataset['Fabric'].isna().sum()

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['Blouse_Fabric'])

fashion_dataset['Fabric'].isna().sum()

pd.unique(fashion_dataset['Better_Cotton'])

sqldf("select Better_Cotton from fashion_dataset where Fabric is null and Better_Cotton in('Better Cotton Initiative','Regular')")

fashion_dataset['New_Cotton'] = fashion_dataset['Better_Cotton'].replace(['Regular', 'Better Cotton Initiative'], ['Cotton',None])

fashion_dataset['Better_Cotton'].isna().sum()

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['New_Cotton'])

fashion_dataset['Fabric'].isna().sum()

pd.unique(fashion_dataset['Dupatta_Fabric'])
fashion_dataset['Dupatta_Fabric'] = fashion_dataset['Dupatta_Fabric'].where(fashion_dataset['Dupatta_Fabric'] != 'NA', None)

sqldf("select Dupatta_Fabric from fashion_dataset where Dupatta_Fabric is not null and Fabric is null")

sqldf("select Weave_Pattern, Fabric from fashion_dataset")

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['Weave_Pattern'])

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['Dupatta_Fabric'])

fashion_dataset['Fabric'].isna().sum()

sqldf("select Saree_Fabric from fashion_dataset where Saree_Fabric is not null and Fabric is null")

fashion_dataset['Fabric'] = fashion_dataset['Fabric'].combine_first(fashion_dataset['Saree_Fabric'])

fashion_dataset['Fabric'].isna().sum()

fashion_dataset['Fabric'].mode()

fashion_dataset.fillna({'Fabric':'Cotton'}, inplace=True)
fashion_dataset['Fabric'].isna().sum()

sqldf("select count(Fabric) from fashion_dataset where Fabric = 'NA'")

### Viewing Weave Pattern, Weave Type & Knit or Woven

pd.unique(fashion_dataset['Knit_or_Woven'])

pd.unique(fashion_dataset['Weave_Type'])

pd.unique(fashion_dataset['Weave_Type'])

pd.unique(fashion_dataset['Weave_Pattern'])

### Formatting Weave Pattern  and Weave Type to fill Knit or Woven

fashion_dataset['Weave_Pattern_2'] = fashion_dataset['Weave_Pattern'].replace(['Regular', 'Jacquard', 'Brocade', 'Dobby', 'Khadi'], ['Woven','Woven','Woven','Woven','Woven'])

fashion_dataset['Weave_Type_2'] = fashion_dataset['Weave_Type'].replace(['Machine Weave',
       'Velvet', 'Denim', 'Handloom', 'Lace', 'Chambray', 'Corduroy'], ['Woven','Woven','Woven','Woven','Knitted','Woven','Woven'])

pd.unique(fashion_dataset['Weave_Pattern_2'])

### Filling Knit_or_Woven

pd.unique(fashion_dataset['Knit_or_Woven'])

fashion_dataset['Knit_or_Woven'].isna().sum()

fashion_dataset['Knit_or_Woven'] = fashion_dataset['Knit_or_Woven'].combine_first(fashion_dataset['Weave_Pattern_2'])

fashion_dataset['Knit_or_Woven'] = fashion_dataset['Knit_or_Woven'].combine_first(fashion_dataset['Weave_Type_2'])

fashion_dataset['Knit_or_Woven'].isna().sum()

pd.unique(fashion_dataset['Knit_or_Woven'])

len(pd.unique(fashion_dataset['Fabric']))

fabrics_knit = sqldf("select distinct Fabric from fashion_dataset where Knit_or_Woven = 'Knitted'")

fabrics_knit.isna().sum()

fabrics_weave = sqldf("select distinct Fabric from fashion_dataset where Knit_or_Woven = 'Woven'")

fabrics_weave.isna().sum()

fabrics_both = sqldf("select distinct Fabric from fashion_dataset where Knit_or_Woven = 'Knitted and Woven'")

fabrics_both.isna().sum()

knit_n_woven = sqldf("select * from fabrics_knit where fabrics_knit.Fabric in(select fabrics_weave.Fabric from fabrics_weave)")

knit_n_woven = sqldf("select * from knit_n_woven where Fabric not in ('Other')")
knit_n_woven

fabrics_both = sqldf("select * from fabrics_both where Fabric not in (select Fabric from knit_n_woven)")

both_fabrics = ps.sqldf("select * from knit_n_woven union all select * from fabrics_both")

both_fabrics.loc[:,'weave'] = 'Knitted and Woven'

fabrics_knit = sqldf("select * from fabrics_knit where Fabric not in (select Fabric from both_fabrics) and Fabric != 'Other'")
fabrics_knit.loc[:,['weave']]='Knitted'

fabrics_knit

fabrics_weave = sqldf("select * from fabrics_weave where Fabric not in (select Fabric from both_fabrics) and Fabric != 'Other'")
fabrics_weave.loc[:,['weave']]='Woven'

fabrics_weave

other_fabrics = sqldf("select Fabric from fashion_dataset where Fabric = 'Other'")

other_woven_fabrics, other_knit_fabrics = train_test_split(other_fabrics, test_size = 0.5,random_state=1)

other_woven_fabrics.loc[:,['weave']]='Woven'

other_knit_fabrics.loc[:,['weave']]= 'Knitted'

other_fabrics = sqldf("select * from other_knit_fabrics union all select * from other_woven_fabrics")

fabric_weave_type = sqldf("select * from fabrics_weave union all select * from fabrics_knit union all select * from both_fabrics union all select * from other_fabrics")
fabric_weave_type

fabric_weave_type.duplicated().sum()

fashion_dataset.info()

fashion_dataset = sqldf("select fashion_dataset.*, fabric_weave_type.weave from fashion_dataset left join fabric_weave_type on (fabric_weave_type.Fabric=fashion_dataset.Fabric)")

fashion_dataset.info()

fashion_dataset.duplicated().sum()

fashion_dataset.drop_duplicates(inplace=True)

fashion_dataset.info()

fashion_dataset['Knit_or_Woven'] = fashion_dataset['Knit_or_Woven'].combine_first(fashion_dataset['weave'])

fashion_dataset['Knit_or_Woven'].isna().sum()

### Filling Print_Pattern_Type null values

pd.unique(fashion_dataset['Print_Pattern_Type'])

fashion_dataset['Print_Pattern_Type'].isna().sum()

sqldf("select Print_Pattern_Type, Pattern, Top_Pattern, Bottom_Pattern from fashion_dataset where Print_Pattern_Type is null and Top_Pattern is not null or Bottom_Pattern is not null or Pattern is not null")

fashion_dataset['Print_Pattern_Type'] = fashion_dataset['Print_Pattern_Type'].combine_first(fashion_dataset['Pattern'])

fashion_dataset['Print_Pattern_Type'] = fashion_dataset['Print_Pattern_Type'].combine_first(fashion_dataset['Bottom_Pattern'])

fashion_dataset['Print_Pattern_Type'] = fashion_dataset['Print_Pattern_Type'].combine_first(fashion_dataset['Top_Pattern'])

fashion_dataset['Print_Pattern_Type'] = fashion_dataset['Print_Pattern_Type'].combine_first(fashion_dataset['Kurta_Pattern'])

fashion_dataset['Print_Pattern_Type'] = fashion_dataset['Print_Pattern_Type'].combine_first(fashion_dataset['Dupatta_Pattern'])

fashion_dataset['Print_Pattern_Type'].isna().sum()

### Filling remaining null values in pattern, weave, and occasion with mode

fashion_dataset['Print_Pattern_Type'].mode()

fashion_dataset['Occasion'].mode()

fashion_dataset['Knit_or_Woven'].mode()

fashion_dataset.fillna({'Print_Pattern_Type':'Solid','Occasion':'Casual','Knit_or_Woven':'Woven'}, inplace=True)
fashion_dataset[['Print_Pattern_Type','Occasion','Knit_or_Woven']].isna().sum()

> Use 'Occasion', 'Print or Pattern Type', 'Pattern', 'Top Pattern', 'Bottom Pattern', 'Dupatta Pattern', 'Kurta Pattern', 'Weave Pattern', 'Knit or Woven', 'Sustainable', 'Sport', 'Fusion_Wear'

fashion_dataset.to_csv('cleaned_1.csv', index=False)
