==> 2. Merging and finishing cleaning.ipynb <==
# importing libraries

import pandas as pd
import numpy as np
import pandasql as ps
from pandasql import sqldf
import ast
import random
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import BayesianRidge
from sklearn import preprocessing
from sklearn import tree
from sklearn.tree import DecisionTreeClassifier
from sklearn import metrics
from sklearn.naive_bayes import GaussianNB
from sklearn.ensemble import RandomForestClassifier

# importing datasets

fashion_dataset= pd.read_csv(r"./cleaned_1.csv")
fashion_dataset

brand_details = pd.read_excel(r".\fashion brand details.xlsx")
brand_details

### Merging the datasets

merged_dataset = ps.sqldf("select fashion_dataset.p_id, fashion_dataset.name, fashion_dataset.price, fashion_dataset.colour, brand_details.*, fashion_dataset.brand, fashion_dataset.ratingCount, fashion_dataset.avg_rating, fashion_dataset.Occasion, fashion_dataset.Print_Pattern_Type, fashion_dataset.Sustainable, fashion_dataset.Knit_or_Woven, fashion_dataset.Fabric, fashion_dataset.Fabric_Purity from fashion_dataset left join brand_details on (fashion_dataset.brand = brand_details.brand_name)")
merged_dataset

### Dropping unnecessary null values

merged_dataset.isna().sum()

sqldf("select * from merged_dataset where p_id is null")

merged_dataset.dropna(subset = "p_id", inplace = True)

sqldf("select * from merged_dataset where name is null")

merged_dataset.dropna(subset = "name", inplace = True)
merged_dataset.isna().sum()

merged_dataset.dropna(subset = "brand", inplace = True)
merged_dataset.isna().sum()

>Finding why there are null brand name and brand ID values

sqldf("select distinct brand from merged_dataset where brand_id is null")

sqldf("select count(distinct brand) from merged_dataset where brand_id is null")

sqldf("select distinct * from merged_dataset where brand is null")

sqldf("select count(distinct brand_id) from merged_dataset")

sqldf("select max(brand_id), brand_name from brand_details")

sqldf("select distinct brand, brand_id from merged_dataset order by brand_id asc")

sqldf("select distinct brand_id from merged_dataset order by brand_id asc")

sqldf("select brand_id from brand_details where brand_name = 'KASSUALLY'")

sqldf("select distinct * from merged_dataset where brand is null")

sqldf("select * from merged_dataset where brand_name is null")

sqldf("select count(distinct brand_id) from merged_dataset")

sqldf("select max(brand_id), brand_name from brand_details")

### Filling brand ID

>Getting rid of null values in brand and brand ID by assigning new IDs

new_id = sqldf("select distinct brand, brand_id from merged_dataset order by brand asc")
new_id

new_id.loc[:, 'brandID'] = range(1, 1021)

merged_dataset = sqldf("select merged_dataset.*, new_id.brandID from merged_dataset left join new_id on (merged_dataset.brand = new_id.brand)")
merged_dataset.drop(['brand_id','brand_name'], axis = 'columns', inplace = True)
merged_dataset.isna().sum()
merged_dataset.info()

merged_dataset.to_csv('cleaned_2.csv', index=False)
