==> 3. Purity and Sustainability.ipynb <==
import pandas as pd
import numpy as np
import pandasql as ps
from pandasql import sqldf
import ast
import random
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split

# importing datasets

merged_dataset= pd.read_csv(r"./cleaned_2.csv")
merged_dataset

merged_dataset.isna().sum()

### Filling Fabric_Purity and Sustainable columns

pd.unique(merged_dataset['Fabric_Purity'])

pd.unique(merged_dataset['Sustainable'])

sqldf('select Fabric_Purity, Sustainable from merged_dataset where Sustainable is not null and Fabric_Purity is not null')

> Creating new column for fabric purity based on sustainable

merged_dataset.loc[:,['pur']]=merged_dataset['Sustainable'].replace(['Regular', 'Sustainable'],['Synthetic','Pure'])

>Filling Fabric Purity with new values

merged_dataset['Fabric_Purity'].isna().sum()

merged_dataset['Fabric_Purity']=merged_dataset['Fabric_Purity'].combine_first(merged_dataset['pur'])

merged_dataset['Fabric_Purity'].isna().sum()

>Creating new column for sustainable values based on  fabric purity values

merged_dataset.loc[:,['sus']]=merged_dataset['Fabric_Purity'].replace(['Blended', 'Pure', 'Synthetic'],['Regular', 'Sustainable','Regular'])

>Filling Sustainable with new values

merged_dataset['Sustainable'].isna().sum()

merged_dataset['Sustainable']=merged_dataset['Sustainable'].combine_first(merged_dataset['sus'])

merged_dataset['Sustainable'].isna().sum()

>Dropping newly created columns

merged_dataset.drop(['pur','sus'], axis='columns', inplace=True)

merged_dataset.isna().sum()

### Filling in remaining null values

print(merged_dataset['colour'].mode())
print(merged_dataset['avg_rating'].mean())
print(merged_dataset['ratingCount'].mean())
print(merged_dataset['Sustainable'].mode())
print(merged_dataset['Fabric_Purity'].mode())
print(merged_dataset['Print_Pattern_Type'].mode())

merged_dataset.fillna({'colour':'Black',"avg_rating":4.1007883731964485, "ratingCount":183.36121433078594,
                       'Sustainable':'Regular','Fabric_Purity':'Synthetic', 
                       'Print_Pattern_Type':'Solid'}, inplace = True)
merged_dataset.isna().sum()

### Converting ratingCount to int

merged_dataset['ratingCount']=merged_dataset['ratingCount'].astype(int)

merged_dataset.info()

### Rounding up avg_rating

merged_dataset['avg_rating']=merged_dataset['avg_rating'].round(1)

merged_dataset['avg_rating']

merged_dataset.to_csv('./new_dataset.csv',index=False)

### Exporting dataset for use in Apache Hive

merged_dataset.to_csv('hive_data.csv', index=False,header=False)

### Exporting dataset for use in Apache MapReduce

final_dataset = merged_dataset.loc[:,['p_id','price','ratingCount','avg_rating','brandID']]

regular_id = sqldf("select brandID from merged_dataset where brand in ('Roadster','Tokyo Talkies','SASSAFRAS','Clora Creation','URBANIC','Mitera','H&M','Anouk','Dupatta Bazaar')")

sustainable_id = sqldf("select brandID from merged_dataset where brand in ('ADA','Taavi','KALINI','Vishudh','Global Desi','DressBerry','anayna','VERO AMORE','UNITED LIBERTY')")

most_sustainable = sqldf("select * from final_dataset where brandID in (select brandID from sustainable_id)")
least_sustainable = sqldf("select * from final_dataset where brandID in (select brandID from regular_id)")

most_sustainable.to_csv('./most_sustainable.csv', index=False, header=False)
least_sustainable.to_csv('./least_sustainable.csv', index=False, header=False)

final_dataset.to_csv('./final_dataset.csv',index=False, header=False)
